% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_table_schema}
\alias{check_table_schema}
\title{Validate table schema}
\usage{
check_table_schema(x, x_schema)
}
\arguments{
\item{x}{data.frame: Table to be validated}

\item{x_schema}{data.frame: Schema the x table must be validated against}
}
\value{
list
}
\description{
Validate table schema
}
\examples{
x <- data.frame(
var1 = c("a", "b"),
var2 = c(1, 2))
x_schema <- data.frame(
var1 = character(0),
var2 = numeric(0))
check_table_schema(x = x, x_schema = x_schema)
}
