% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk_out_gdp.R
\name{chk_out_gdp}
\alias{chk_out_gdp}
\title{Implement validations for the auxiliary GDP table}
\usage{
chk_out_gdp(
  x,
  schema,
  max_year_to_year_change = 1.9,
  distinct_rows = c("country_code", "data_level", "year")
)
}
\arguments{
\item{x}{data.frame: GDP table}

\item{schema}{data.frame: Expected schema for GDP table}

\item{max_year_to_year_change}{numeric: Maximum expected value for year-to-year
change in GDP}

\item{distinct_rows}{character: Combination of columns for which there should
be no duplicated rows}
}
\value{
pointblank object
}
\description{
Implement validations for the auxiliary GDP table
}
