% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk_out_prod_svy_estimation.R
\name{chk_out_prod_svy_estimation}
\alias{chk_out_prod_svy_estimation}
\title{Implement validations for the prod_svy_estimation table}
\usage{
chk_out_prod_svy_estimation(
  x,
  schema,
  distinct_rows = c("country_code", "reporting_year", "welfare_type",
    "reporting_level")
)
}
\arguments{
\item{x}{data.frame: prod_svy_estimation table}

\item{schema}{data.frame: Expected schema for prod_svy_estimation table}

\item{distinct_rows}{character: Combination of columns for which there should
be no duplicated rows}
}
\value{
pointblank object
}
\description{
Implement validations for the prod_svy_estimation table
}
